srd: Semantic Adjustments
Previous code was not implemented accordingly with the desired
semantic.
We believe the current code states better the purpouse, specifically
about the concept of metric. There are things that may change in the
future, like the metric relantionship with degrader and compressor,
but we will keep it as it is for the time being.

New files:
- src/compressor.py: Encapsulates what a compressor function should do
- src/degrader.py: Classes for applying the degradations randomly

Modified files:
- src/algorithm.py: Much of the code moved to src/degrader
- src/data/util.py: added df_to_str function
- src/text/tokenizer: added lines function
- src/metric.py: A metric needs a degrader and a compression function
- src/experiments.py: Adjustments for the new code